dom/plugins/test/mochitest/test_twostreams.html

<html>
<head>
  <title>Dual NPAPI NP_ASFILEONLY NPStream Test</title>
  <script type="text/javascript" 
          src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="plugin-utils.js"></script>
  <link rel="stylesheet" type="text/css" 
        href="/tests/SimpleTest/test.css" />
</head>
<body>
  <p id="display"></p>

  <script type="text/javascript">
  SimpleTest.expectAssertions(0, 2);

  SimpleTest.waitForExplicitFinish();
  setTestPluginEnabledState(SpecialPowers.Ci.nsIPluginTag.STATE_ENABLED);

  var framesToLoad = 2;
  function frameLoaded(id) {
    var frame = document.getElementById('testframe' + id);
    // We have to use SpecialPowers because nptest.cpp prepends
    // data: whichs makes the frame cross origin with the including page.
    var wrappedDoc = SpecialPowers.wrap(frame).contentDocument;

    if (!wrappedDoc.body.innerHTML.length)
      return;

    --framesToLoad;
    if (0 == framesToLoad) {
      var frame1 = document.getElementById('testframe1');
      var frame2 = document.getElementById('testframe2');
      var wrappedDocFrame1 = SpecialPowers.wrap(frame1).contentDocument;
      var wrappedDocFrame2 = SpecialPowers.wrap(frame2).contentDocument;

      is(wrappedDocFrame1.body.innerHTML, wrappedDocFrame2.body.innerHTML,
         "Frame contents should match");
      SimpleTest.finish();
    }
  }
  </script>

  <iframe id="testframe1" name="testframe1" onload="frameLoaded(1)"></iframe>
  <iframe id="testframe2" name="testframe2" onload="frameLoaded(2)"></iframe>

  <embed src="loremipsum_nocache.txt" streammode="asfileonly"
         frame="testframe1"
         id="embedtest" style="width: 400px; height: 100px;"
         type="application/x-test"></embed>
  <embed src="loremipsum_nocache.txt" streammode="asfileonly"
	 frame="testframe2"
	 id="embedtest2" style="width: 400px; height: 100px;"
	 type="application/x-test"></embed>
